refactor: move logout methods into a trait
